Try rearranging conversion methods
